+++ R/tsea_mabs.R
@@ -50,6 +50,9 @@
 #' complete drug-target annotations. Choosing a single
 #' annotation source results in sparser drug-target annotations 
 #' (particularly CLUE), and thus less complete enrichment results.
+#' @param readable TRUE or FALSE, it applies when type is `KEGG` or `Reactome`
+#' indicating whether to convert gene Entrez ids to gene Symbols in the 'itemID' 
+#' column in the result table.
 #' @return \code{\link{feaResult}} object, the result table contains the
 #' enriched functional categories (e.g. GO terms or KEGG pathways) ranked by 
 #' the corresponding enrichment statistic.
@@ -77,7 +80,7 @@ tsea_mabs <- function(drugs,
                       nPerm=1000,  
                       pAdjustMethod="BH", pvalueCutoff=0.05,
                       minGSSize=5, maxGSSize=500, 
-                      dt_anno="all"){
+                      dt_anno="all", readable=FALSE){
   if(!any(type %in% c("GO", "KEGG", "Reactome"))){
     stop('"type" argument needs to be one of "GO", "KEGG" or "Reactome"')
   }
@@ -131,7 +134,8 @@ tsea_mabs <- function(drugs,
     mabs_res <- mabsKEGG(geneList=tar_total_weight, organism='hsa', 
                        keyType='kegg', nPerm = nPerm, 
                        minGSSize = minGSSize, maxGSSize=maxGSSize, 
-                       pvalueCutoff=pvalueCutoff, pAdjustMethod = pAdjustMethod)
+                       pvalueCutoff=pvalueCutoff, pAdjustMethod = pAdjustMethod,
+                       readable=readable)
   }
   
   if(type=="Reactome"){
@@ -148,7 +152,8 @@ tsea_mabs <- function(drugs,
     mabs_res <- mabsReactome(geneList=tar_total_weight, organism='human', 
                              nPerm = nPerm, 
                              minGSSize = minGSSize, maxGSSize=maxGSSize, 
-                             pvalueCutoff=pvalueCutoff, pAdjustMethod=pAdjustMethod)
+                             pvalueCutoff=pvalueCutoff, pAdjustMethod=pAdjustMethod,
+                             readable=readable)
   }
   
   if(is.null(mabs_res))